test: move controlplaneloadbalancer unit tests to its own package
